refactor: organize imports feat: add support for DarkPeers
