+++ src/backend/trackers/__init__.py
@@ -1,18 +1,23 @@
-from src.backend.trackers.morethantv import MTVUploader, MTVSearch, mtv_uploader
-from src.backend.trackers.torrentleech import TLUploader, TLSearch, tl_upload
-from src.backend.trackers.beyondhd import BHDUploader, BHDSearch, bhd_uploader
-from src.backend.trackers.passthepopcorn import ptp_uploader, PTPSearch, PTPUploader
+from src.backend.trackers.aither import AitherSearch, AitherUploader, aither_uploader
+from src.backend.trackers.beyondhd import BHDSearch, BHDUploader, bhd_uploader
+from src.backend.trackers.darkpeers import (
+    DarkPeersSearch,
+    DarkPeersUploader,
+    dp_uploader,
+)
+from src.backend.trackers.huno import HunoSearch, HunoUploader, huno_uploader
+from src.backend.trackers.lst import LSTSearch, LSTUploader, lst_uploader
+from src.backend.trackers.morethantv import MTVSearch, MTVUploader, mtv_uploader
+from src.backend.trackers.passthepopcorn import PTPSearch, PTPUploader, ptp_uploader
+from src.backend.trackers.reelflix import ReelFlixSearch, ReelFlixUploader, rf_uploader
+from src.backend.trackers.torrentleech import TLSearch, TLUploader, tl_upload
 from src.backend.trackers.unit3d_base import (
-    Unit3dBaseSearch,
-    Unit3dBaseUploader,
     CategoryEnums,
     ResolutionEnums,
     TypeEnums,
+    Unit3dBaseSearch,
+    Unit3dBaseUploader,
 )
-from src.backend.trackers.reelflix import ReelFlixSearch, rf_uploader, ReelFlixUploader
-from src.backend.trackers.aither import AitherSearch, aither_uploader, AitherUploader
-from src.backend.trackers.huno import HunoSearch, huno_uploader, HunoUploader
-from src.backend.trackers.lst import LSTSearch, lst_uploader, LSTUploader
 
 __all__ = (
     "MTVUploader",
@@ -44,4 +49,7 @@ __all__ = (
     "LSTSearch",
     "lst_uploader",
     "LSTUploader",
+    "DarkPeersSearch",
+    "dp_uploader",
+    "DarkPeersUploader",
 )
